Compile of Header module
